Enabled counters to hold BaseValue separately in HKeyPerfDataReader
This also fixes the wrong value of `windows_os_paging_free_bytes`.

This changes the output data shape of the reader.
The value of `Counters` hash of `PerfInstance` is `PerfCounter` now,
so we have to add `.value` or `.base_value` to take the value.

The previous implementation was not able to handle the case of
two counters with the same name, which occurs with certain counter types.
This made the value of `windows_os_paging_free_bytes` wrong.
I confirmed that the other values are not affected by this problem.

This fix solves the problem by handling `base_value` separately from `value`.

For counters that have both values, we need to consider which value to use.

About `BaseValue` and `CounterType`:
* https://docs.microsoft.com/en-us/previous-versions/windows/it-pro/windows-server-2003/cc785636(v=ws.10)
* https://docs.microsoft.com/en-us/dotnet/api/system.diagnostics.performancecountertype

Signed-off-by: daipom <fukuda@clear-code.com>